Crate mba

Mixed Boolean-Arithmetic expressions and operations on them.

Modules

bitwise_expr

This module contains bitwise expressions (BExpr), which are just bitwise expressions on n bits, and linear combinations of bitwise expressions (LBExpr).

expr

General expressions.

formatter

Formatting for expressions.

lattice

Integer lattices and algorithms.

linear_mba

Linear Mixed Boolean-Arithmetic.

matrix

Matrices. See the module documentation for crate::vector. The same thing applies here.

perm_poly

Binary permutation polynomials.

poly

Polynomials.

rings

A trait for the kind of number (ring) we are working with. These traits exist and look the way they do for multiple reasons:

simplify_boolean

Simplify boolean expressions using egg. There are some obvious improvements you could make:

solver

Solves systems of linear equations modulo some number, usually a power of two.

tex

Generates LaTeX code for various objects.

valuation

Basically a key-value store for variable names and their values, but you can specify what to do when a variable is not found.

vector

Vectors.

Macros

choose_binary_ring

Choose the most efficient BinaryRing for the given number of bits.

Structs

Symbol

A interned string in the global symbol table.