mba

Module bitwise_expr

This module contains bitwise expressions (BExpr), which are just bitwise expressions on n bits, and linear combinations of bitwise expressions (LBExpr).

Structs

LBExpr

LBExpr is short for “Linear combination of Bitwise Expressions” These are the expressions for which rewrite rules can be efficiently generated.

Enums

BExpr

Represents an expression whose value bit i depends only on the values of bits i of the variables.