mba::bitwise_expr

Struct LBExpr

pub struct LBExpr<R: Ring>(pub Vec<(R::Element, BExpr)>);

LBExpr is short for “Linear combination of Bitwise Expressions” These are the expressions for which rewrite rules can be efficiently generated.

Tuple Fields
0: Vec<(R::Element, BExpr)>
Implementations

impl<R: Ring> LBExpr<R>

pub fn zero() -> Self

Create the empty linear combination. It evaluates to 0.

pub fn constant(c: R::Element, r: &R) -> Self

Creates an expression that equals a constant.

pub fn var<T: Into<Symbol>>(name: T) -> Self

Creates an expression that equals a variable.

pub fn remove_zero_terms(&mut self)

Removes all terms with coefficient 0.

pub fn vars(&self) -> Vec<Symbol>

Returns all variables in the expression. This will include duplicates.

pub fn eval(&self, v: &mut Valuation<R>, r: &R) -> R::Element
where R: BinaryRing,

Evaluate an expression with a valuation for the occurring variables.

pub fn complexity(&self) -> u32
where R: BinaryRing,

Returns a measure of the complexity of the expression.

pub fn from_string(s: String, r: &R) -> Result<Self, String>

Parse a string to an expression. Note that this function is extremely limited and expects very specific syntax. It is used for convenience when testing things and not really meant to be used by something outside this crate.

pub fn to_expr(&self, r: &R) -> ExprOp<R>

Converts an LBExpr to an ExprOp.

impl<R: Ring> LBExpr<R>

pub fn display<'a>(&'a self, formatter: Formatter) -> LBExprFormatter<'a, R>

Creates a wrapper struct that implements Display and uses the given formatter to format the expression.

pub fn display_function<'a>( &'a self, formatter: Formatter, function_name: &'a str, r: &'a R, ) -> FunctionFormatter<'a, R, LBExprFormatter<'a, R>>

Creates a wrapper struct that implements Display and uses the given formatter to format the expression as a function with the given function_name. This doesn’t make sense for Formatter::Tex so it is the same as as Self::display in that case, except that it is slower because it computes the list of variables on construction.