mba::expr

Struct Expr

pub struct Expr<R: Ring>(/* private fields */);

Implementations

impl<R: Ring> Expr<R>

pub fn new(e: ExprOp<R>) -> Self

pub fn as_ref(&self) -> &ExprOp<R>

pub fn as_ptr(&self) -> *const ExprOp<R>

pub fn strong_count(&self) -> usize

pub fn ptr_eq(&self, other: &Self) -> bool

pub fn simplify(&mut self)

Simplify the expression. This does only trivial simplifications. E.g. multiplication by one or zero.

pub fn eval(&self, v: &mut Valuation<R>, r: &R) -> R::Element
where R: BinaryRing,

Evaluate an expression.

pub fn substitute(&mut self, var: Symbol, s: &mut Expr<R>)

Substitutes an expression for a variable. If the Rc’s in this expression are shared with other expressions then this will also substitute in those.

pub fn deep_copy(&self) -> Self

Creates a copy of the expression where no subexpression is shared among the copies.

pub fn from_string(s: String, r: &R) -> Result<Expr<R>, String>

Parse an expression from a string.

impl<R: Ring> Expr<R>

pub fn display<'a>( &'a self, formatter: Formatter, tabs: usize, r: &'a R, ) -> ExprFormatter<'a, R>

pub fn display_function<'a>( &'a self, formatter: Formatter, function_name: &'a str, r: &'a R, ) -> FunctionFormatter<'a, R, ExprFormatter<'a, R>>