mba::lattice

Function cvp_planes

pub fn cvp_planes<R: Ring, W: WorkingTypeFor<R>>(
    basis: &OwnedMatrix<R>,
    t: &VectorView<R>,
    rad_sqr: Option<W::Element>,
    wt: &W,
    ring: &R,
) -> Option<OwnedVector<R>>

Solves the CVP exactly.

basis is the matrix that contains the basis as rows.

t is the target vector.

prec is the precision of the floating point numbers used.

rad_sqr is an optional float that contains the square of the maximum distance to search for.

The returned vector is the vector of coefficients. This will always return some vector unless no vector is within r of the target. This algorithm is a generalization Babai’s nearest plane algorithm that searches all planes that could contain the closest vector. It is the simplest one I could think of.