mba::lattice

Struct Lattice

pub struct Lattice<R: Ring> {
    pub basis: OwnedMatrix<R>,
}

A “lattice”, i.e. the span of a set of vectors. This can never be empty because it always contains the zero vector even if Lattice::basis has zero rows. The matrix always stores the ambient dimension (the number of columns) even if there are zero rows.

Fields
basis: OwnedMatrix<R>

The basis matrix of the lattice. The basis vectors are the rows of the matrix.

Implementations

impl<R: Ring> Lattice<R>

pub fn zero(ambient_dim: usize) -> Self

Creates a lattice that contains only the zero vector.

pub fn from_basis(basis: OwnedMatrix<R>) -> Self

The lattice basis are the rows of the matrix.

pub fn from_generating_set(generating_set: OwnedMatrix<R>, r: &R) -> Self
where R: IntDivRing,

The rows of the matrix generate the lattice but are potentially linearly dependent. This function will compute the Hermite normal form and remove zero rows.

pub fn rank(&self) -> usize

Returns the rank of the lattice, i.e. the number if basis vectors.

pub fn ambient_dim(&self) -> usize

Returns the dimension of the ambient space or 0 if the lattice is empty.

pub fn at<S: VectorStorage<R> + ?Sized>( &self, coefficients: &Vector<R, S>, r: &R, ) -> OwnedVector<R>

Returns the vector on the lattice that is the linear combination of the basis vectors with the given coefficients.

pub fn sample_point<Rand: Rng>(&self, rng: &mut Rand, r: &R) -> OwnedVector<R>

Returns a random point on the lattice.

pub fn canonicalize(&self, ring: &R) -> Lattice<Z>
where R: BinaryRing,

Returns an integer lattice such that two lattices are equal (i.e. contain the same points) iff l1.canonicalize() == l2.canonicalize().

The basis of the returned lattice is guaranteed to be in Hermite normal form and of full rank. In particular, the diagonal entries are positive.

pub fn size_reduce(&mut self, r: &R)
where R: IntDivRing,

Size reduce the basis. This essentially is Gram-Schmidt but rounding the coefficients to integers.

pub fn lll<W: WorkingTypeFor<R>>(&mut self, delta: &W::Element, wt: &W, r: &R)
where R: IntDivRing,

Performs LLL basis reduction.

pub fn cvp_planes_coeff<W: WorkingTypeFor<R>>( &self, t: &VectorView<R>, rad_sqr: Option<W::Element>, wt: &W, r: &R, ) -> Option<OwnedVector<R>>

pub fn cvp_planes<W: WorkingTypeFor<R>>( &self, t: &VectorView<R>, rad_sqr: Option<W::Element>, wt: &W, r: &R, ) -> Option<OwnedVector<R>>

pub fn cvp_rounding_coeff<W: WorkingTypeFor<R>>( &self, t: &VectorView<R>, wt: &W, r: &R, ) -> OwnedVector<R>

pub fn cvp_rounding<W: WorkingTypeFor<R>>( &self, t: &VectorView<R>, wt: &W, r: &R, ) -> OwnedVector<R>

pub fn cvp_nearest_plane_coeff<W: WorkingTypeFor<R>>( &self, t: &VectorView<R>, wt: &W, r: &R, ) -> OwnedVector<R>

pub fn cvp_nearest_plane<W: WorkingTypeFor<R>>( &self, t: &VectorView<R>, wt: &W, r: &R, ) -> OwnedVector<R>

impl<R: Ring> Lattice<R>

pub fn to_tex(&self) -> TexLattice<'_, R>