mba

Module linear_mba

Linear Mixed Boolean-Arithmetic.

Structs

DeobfuscationConfig

Configuration for deobfuscation.

ObfuscationConfig

More obscure settings for internals of the obfuscation. Use ObfuscationConfig::default() to get a reasonable default.

Subs

A list of substitutions. See e.g. expr_to_bexpr on what this is used for.

Enums

RewriteTries

Used internally in ObfuscationConfig.

SolutionAlgorithm

The algorithm to use for finding a small solution.

Functions

collect_solution

Converts a solution to the linear system and the rewrite operations to a single LBExpr with no duplicate crate::bitwise_expr::BExprs.

deobfuscate

deobfuscate_lbexpr

Deobfuscate a linear MBA expression.

expr_to_bexpr

Converts part of an expression to a BExpr, such that if you substituted the Exprs in subs for the variables, you would get the original Expr.

expr_to_lbexpr

Converts part of an expression to an LBExpr.

is_linear_mba

Checks if an Expr implements a linear MBA function by checking if the fundamental theorem of linear MBA holds for all inputs.

is_probably_linear_mba

Checks if an Expr implements a linear MBA function by randomly sampling num_inputs inputs and checking if the fundamental theorem of linear MBA holds for those inputs.

obfuscate

Obfuscate any expression using linear MBA.

rewrite

Rewrite an LBExpr using a set of LBExprs.

solve_linear_system

Solves the linear system for rewriting expr using the rewrite operations ops.