mba::linear_mba

Function expr_to_bexpr

pub fn expr_to_bexpr<R: BinaryRing>(
    e: &Expr<R>,
    subs: &mut Subs<R>,
    force: bool,
) -> Option<BExpr>

Converts part of an expression to a BExpr, such that if you substituted the Exprs in subs for the variables, you would get the original Expr.

It will generally try to make the LBExpr as big as possible.

If force is false, it will return None if the top-most operation is not a BExpr operation. Otherwise, it will return a BExpr::Var whose substitution is the original expression.