mba::linear_mba

Function expr_to_lbexpr

pub fn expr_to_lbexpr<R: BinaryRing>(
    e: &Expr<R>,
    force: bool,
    r: &R,
) -> Option<(LBExpr<R>, Subs<R>)>

Converts part of an expression to an LBExpr.

Returns the LBExpr and a list of substitutions, such that substituting the expressions in the list into the variables, would give the original expression.

It will generally try to make the LBExpr as big as possible.

If force is false, it will return None if the top-most operation is not something a LBExpr can represent. Otherwise, it will return a new variable, whose substitution is the original expression.