mba

Module matrix

Matrices. See the module documentation for crate::vector. The same thing applies here.

Structs

ColumnVectorStorage

Matrix

A matrix.

MatrixSliceIter

Iterate over vectors of the matrix that are contiguous in memory. This is usually rows for OwnedMatrix and MatrixView and columns for TransposedMatrixView.

MatrixSliceIterMut

Mutable version of MatrixSliceIter.

MatrixStridedIter

Iterate over vectors of the matrix that are strided in memory. This is usually columns for OwnedMatrix and MatrixView and rows for TransposedMatrixView.

MatrixStridedIterMut

See MatrixStridedIter.

OwnedMatrixStorage

RowVectorStorage

SliceMatrixStorage

Very hacky, see StrideStorage for explanation. Metadata stores the number of rows (in the least significant 4 bytes) and columns (in the most significant 4 bytes).

TransposedMatrixStorage

Very hacky, see StrideStorage for explanation. Metadata stores the number of columns (in the least significant 4 bytes) and rows (in the most significant 4 bytes). This allows you to transmute a SliceMatrixStorage reference into a TransposedMatrixStorage reference while transposing the view.

Traits

MatrixStorage

How are the entries of a matrix stored?

Type Aliases

ColumnVector

A matrix with a single column.

MatrixView

OwnedMatrix

RowVector

A matrix with a single row.

TransposedMatrixView