mba::matrix

Struct Matrix

pub struct Matrix<R: Ring, S: MatrixStorage<R> + ?Sized> { /* private fields */ }

A matrix.

Implementations

impl<R: Ring, S: MatrixStorage<R> + ?Sized> Matrix<R, S>

pub fn num_rows(&self) -> usize

The number of rows of the matrix.

pub fn num_cols(&self) -> usize

The number of columns of the matrix.

pub fn min_dim(&self) -> usize

Returns the smaller of the two dimensions.

pub fn is_empty(&self) -> bool

Is the matrix empty, i.e. has it zero rows or columns?

pub fn row(&self, r: usize) -> &Vector<R, S::RowVecStorage>

Returns a view of the row r.

pub fn row_mut(&mut self, r: usize) -> &mut Vector<R, S::RowVecStorage>

Returns a mutable view of the row r.

pub fn col(&self, c: usize) -> &Vector<R, S::ColVecStorage>

Returns a view of the column c.

pub fn col_mut(&mut self, c: usize) -> &mut Vector<R, S::ColVecStorage>

Returns a mutable view of the column c.

pub fn entry(&self, r: usize, c: usize) -> &R::Element

Returns a reference to the entry at row r and column c.

pub fn entry_mut(&mut self, r: usize, c: usize) -> &mut R::Element

Returns a mutable reference to the entry at row r and column c.

pub fn rows(&self) -> S::RowIter<'_>

Returns an iterator over the rows.

pub fn rows_mut(&mut self) -> S::RowIterMut<'_>

Returns an iterator over the mutable rows.

pub fn cols(&self) -> S::ColIter<'_>

Returns an iterator over the columns.

pub fn cols_mut(&mut self) -> S::ColIterMut<'_>

Returns an iterator over the mutable columns.

pub fn entries_row_major(&self) -> impl Iterator<Item = &R::Element>

Returns an iterator over the entries in row-major order.

pub fn entries_row_major_mut(&mut self) -> impl Iterator<Item = &mut R::Element>

Returns an iterator over the mutable entries in row-major order.

pub fn entries_col_major(&self) -> impl Iterator<Item = &R::Element>

Returns an iterator over the entries in column-major order.

pub fn entries_col_major_mut(&mut self) -> impl Iterator<Item = &mut R::Element>

Returns an iterator over the mutable entries in column-major order.

pub fn transform<U: Ring, F: FnMut(&R::Element) -> U::Element>( &self, f: F, ) -> OwnedMatrix<U>

Call a function on each entry and return the resulting matrix.

pub fn map_mut<F: FnMut(&mut R::Element)>(&mut self, f: F)

Call a function on each entry.

pub fn swap_rows(&mut self, i: usize, j: usize)

Swap two rows.

pub fn swap_columns(&mut self, i: usize, j: usize)

Swap two columns.

pub fn get_rows_mut( &mut self, i: usize, j: usize, ) -> (&mut Vector<R, S::RowVecStorage>, &mut Vector<R, S::RowVecStorage>)

Returns a mutable references to two rows of the matrix. Panics if the indices are the same.

pub fn to_owned(&self) -> OwnedMatrix<R>

Returns a copy of the matrix.

pub fn transposed(&self) -> OwnedMatrix<R>

Creates an owned matrix that is the transpose of the current matrix.

pub fn negate_row(&mut self, row: usize, r: &R)

Negates all elements of a row.

pub fn negate_col(&mut self, col: usize, r: &R)

Negates all elements of a column.

pub fn row_multiply(&mut self, row: usize, c: &R::Element, r: &R)

Multiply a row by an element. Note that this is only an invertible operation if c is a unit.

pub fn col_multiply(&mut self, col: usize, c: &R::Element, r: &R)

Multiply a column by an element. Note that this is only an invertible operation if c is a unit.

pub fn row_multiply_add(&mut self, n: usize, m: usize, c: &R::Element, r: &R)

Add a scaled row to another row. N += M * c. m and n can not be equal.

pub fn col_multiply_add(&mut self, n: usize, m: usize, c: &R::Element, r: &R)

Add a scaled column to another column. N += M * c. m and n can not be equal.

pub fn mul<T>(&self, rhs: &Matrix<R, T>, ring: &R) -> OwnedMatrix<R>
where T: MatrixStorage<R> + ?Sized,

Multiply two matrices.

pub fn mul_vec_post<T>(&self, rhs: &Vector<R, T>, r: &R) -> OwnedVector<R>
where T: VectorStorage<R> + ?Sized,

Post-multiply a matrix with a vector.

pub fn mul_vec_pre<T>(&self, lhs: &Vector<R, T>, r: &R) -> OwnedVector<R>
where T: VectorStorage<R> + ?Sized,

Pre-multiply a matrix with a vector.

pub fn print_rows(&self)

Prints each row of the matrix in its own line using the debug formatter.

impl<R: Ring> Matrix<R, OwnedMatrixStorage<R>>

pub fn empty() -> Self

Return an empty matrix.

pub fn num_entries(&self) -> usize

Returns the number of entries.

This function exists because OwnedMatrix stores the number of entries directly, so it doesn’t have to be computed.

pub fn as_slice(&self) -> &[R::Element]

Returns a slice containing all elements.

pub fn as_slice_mut(&mut self) -> &mut [R::Element]

Returns a slice containing all elements.

pub fn view(&self) -> &MatrixView<R>

Returns a view of this matrix.

pub fn view_mut(&mut self) -> &mut MatrixView<R>

Returns a mutable view of this matrix.

pub fn into_vector(self) -> OwnedVector<R>

Turns this matrix into a vector of dimension r*c.

pub fn from_array<U: Into<R::Element>, const RS: usize, const CS: usize>( a: [[U; CS]; RS], ) -> Self

Creates a matrix from an array of rows.

pub fn from_iter<I: Iterator<Item = R::Element>>( r: usize, c: usize, iter: I, ) -> Self

Creates a matrix from an iterator in row-major order.

pub fn try_from_iter<E, I: Iterator<Item = Result<R::Element, E>>>( r: usize, c: usize, iter: I, ) -> Result<Self, E>

Tries to create a matrix from an iterator in row-major order.

pub fn from_rows<U, V>(rows: &[U]) -> Self
where U: AsRef<[V]>, V: Into<R::Element> + Clone,

Creates a matrix from slice of rows.

pub fn random<Rand: Rng>(r: usize, c: usize, ring: &R, rng: &mut Rand) -> Self

Creates a random matrix.

pub fn transpose(&self) -> &TransposedMatrixView<R>

View of the transpose.

pub fn transpose_mut(&mut self) -> &mut TransposedMatrixView<R>

Mutable view of the transpose.

pub fn shrink(&mut self, new_nrows: usize)

Removes rows at the end of the matrix.

pub fn zero(r: usize, c: usize) -> Self

Returns an r×c zero matrix.

pub fn identity(n: usize) -> Self

Returns an nxn identity matrix.

pub fn remove_zero_rows(&mut self)

Remove rows of zeros at the end of the matrix.

pub fn append_zero_rows(&mut self, n: usize)

Append n zero rows to the end of the matrix.

impl<R: Ring> Matrix<R, SliceMatrixStorage<R>>

pub fn transpose(&self) -> &TransposedMatrixView<R>

Returns a view of the matrix transposed.

pub fn transpose_mut(&mut self) -> &mut TransposedMatrixView<R>

Returns a mutable view of the matrix transposed.

impl<R: Ring> Matrix<R, TransposedMatrixStorage<R>>

pub fn transpose(&self) -> &MatrixView<R>

Returns a view of the matrix transposed.

pub fn transpose_mut(&mut self) -> &mut MatrixView<R>

Returns a mutable view of the matrix transposed.

impl<R: Ring, S: MatrixStorage<R> + ?Sized> Matrix<R, S>

pub fn to_tex(&self) -> TexMatrix<'_, R, S>