mba::matrix

Trait MatrixStorage

pub trait MatrixStorage<R: Ring>: 'static {
    type RowVecStorage: VectorStorage<R> + ?Sized;
    type ColVecStorage: VectorStorage<R> + ?Sized;
    type RowIter<'a>: DoubleEndedIterator<Item = &'a Vector<R, Self::RowVecStorage>>;
    type RowIterMut<'a>: DoubleEndedIterator<Item = &'a mut Vector<R, Self::RowVecStorage>>;
    type ColIter<'a>: DoubleEndedIterator<Item = &'a Vector<R, Self::ColVecStorage>>;
    type ColIterMut<'a>: DoubleEndedIterator<Item = &'a mut Vector<R, Self::ColVecStorage>>;

    // Required methods
    fn rows(&self) -> usize;
    fn cols(&self) -> usize;
    fn row(&self, r: usize) -> &Self::RowVecStorage;
    fn row_mut(&mut self, r: usize) -> &mut Self::RowVecStorage;
    fn row_iter<'a>(&'a self) -> Self::RowIter<'a>;
    fn row_iter_mut<'a>(&'a mut self) -> Self::RowIterMut<'a>;
    fn col(&self, c: usize) -> &Self::ColVecStorage;
    fn col_mut(&mut self, c: usize) -> &mut Self::ColVecStorage;
    fn col_iter<'a>(&'a self) -> Self::ColIter<'a>;
    fn col_iter_mut<'a>(&'a mut self) -> Self::ColIterMut<'a>;

    // Provided methods
    fn entry(&self, r: usize, c: usize) -> &R::Element { ... }
    fn entry_mut(&mut self, r: usize, c: usize) -> &mut R::Element { ... }
}

How are the entries of a matrix stored?

Required Associated Types

type RowVecStorage: VectorStorage<R> + ?Sized

Row/Column vector type.

type ColVecStorage: VectorStorage<R> + ?Sized

type RowIter<'a>: DoubleEndedIterator<Item = &'a Vector<R, Self::RowVecStorage>>

Row/column iter type.

type RowIterMut<'a>: DoubleEndedIterator<Item = &'a mut Vector<R, Self::RowVecStorage>>

type ColIter<'a>: DoubleEndedIterator<Item = &'a Vector<R, Self::ColVecStorage>>

type ColIterMut<'a>: DoubleEndedIterator<Item = &'a mut Vector<R, Self::ColVecStorage>>

Required Methods

fn rows(&self) -> usize

Returns the number of rows.

fn cols(&self) -> usize

Returns the number of columns.

fn row(&self, r: usize) -> &Self::RowVecStorage

Returns a view of the row r.

fn row_mut(&mut self, r: usize) -> &mut Self::RowVecStorage

Returns a mutable view of the row r.

fn row_iter<'a>(&'a self) -> Self::RowIter<'a>

Returns an iterator over the rows.

fn row_iter_mut<'a>(&'a mut self) -> Self::RowIterMut<'a>

Returns an iterator over the rows.

fn col(&self, c: usize) -> &Self::ColVecStorage

Returns a view of the column c.

fn col_mut(&mut self, c: usize) -> &mut Self::ColVecStorage

Returns a mutable view of the column c.

fn col_iter<'a>(&'a self) -> Self::ColIter<'a>

Returns an iterator over the columns.

fn col_iter_mut<'a>(&'a mut self) -> Self::ColIterMut<'a>

Returns an iterator over the columns.

Provided Methods

fn entry(&self, r: usize, c: usize) -> &R::Element

Returns a reference to the entry at row r and column c.

fn entry_mut(&mut self, r: usize, c: usize) -> &mut R::Element

Returns a mutable reference to the entry at row r and column c.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.