mba::matrix

Type Alias OwnedMatrix

pub type OwnedMatrix<R> = Matrix<R, OwnedMatrixStorage<R>>;

Aliased Type

pub struct OwnedMatrix<R> { /* private fields */ }

Implementations

impl<R: Ring> OwnedMatrix<R>

pub fn empty() -> Self

Return an empty matrix.

pub fn num_entries(&self) -> usize

Returns the number of entries.

This function exists because OwnedMatrix stores the number of entries directly, so it doesn’t have to be computed.

pub fn as_slice(&self) -> &[R::Element]

Returns a slice containing all elements.

pub fn as_slice_mut(&mut self) -> &mut [R::Element]

Returns a slice containing all elements.

pub fn view(&self) -> &MatrixView<R>

Returns a view of this matrix.

pub fn view_mut(&mut self) -> &mut MatrixView<R>

Returns a mutable view of this matrix.

pub fn into_vector(self) -> OwnedVector<R>

Turns this matrix into a vector of dimension r*c.

pub fn from_array<U: Into<R::Element>, const RS: usize, const CS: usize>( a: [[U; CS]; RS], ) -> Self

Creates a matrix from an array of rows.

pub fn from_iter<I: Iterator<Item = R::Element>>( r: usize, c: usize, iter: I, ) -> Self

Creates a matrix from an iterator in row-major order.

pub fn try_from_iter<E, I: Iterator<Item = Result<R::Element, E>>>( r: usize, c: usize, iter: I, ) -> Result<Self, E>

Tries to create a matrix from an iterator in row-major order.

pub fn from_rows<U, V>(rows: &[U]) -> Self
where U: AsRef<[V]>, V: Into<R::Element> + Clone,

Creates a matrix from slice of rows.

pub fn random<Rand: Rng>(r: usize, c: usize, ring: &R, rng: &mut Rand) -> Self

Creates a random matrix.

pub fn transpose(&self) -> &TransposedMatrixView<R>

View of the transpose.

pub fn transpose_mut(&mut self) -> &mut TransposedMatrixView<R>

Mutable view of the transpose.

pub fn shrink(&mut self, new_nrows: usize)

Removes rows at the end of the matrix.

pub fn zero(r: usize, c: usize) -> Self

Returns an r×c zero matrix.

pub fn identity(n: usize) -> Self

Returns an nxn identity matrix.

pub fn remove_zero_rows(&mut self)

Remove rows of zeros at the end of the matrix.

pub fn append_zero_rows(&mut self, n: usize)

Append n zero rows to the end of the matrix.