mba

Module perm_poly

Binary permutation polynomials.

Structs

ZeroIdeal

The zero ideal is the ideal of all polynomials that evaluate to zero everywhere in the polynomial ring Z_{2^n}[X].

Functions

compose

Computes the composition p(q(x)) of two polynomials. It does not simplify the result as much as possible, so it is recommended to call Poly::simplify afterwards. It does however reduce the degree of the result.

compute_inverse

Computes the inverse of a permutation polynomial using Newton’s Method.

compute_inverse_generator

Computes the inverse of a permutation polynomial by using f as a generator.

compute_inverse_interpolation

Compute the inverse using interpolation.

is_perm_poly

Is this function a permutation polynomial?

perm_pair

Returns a pair of permutation polynomials mod 2^n. The functions are inverses of each other. degree specifies the degree of the first polynomial.

random_perm_poly

Returns a random permutation polynomial.