mba::poly

Struct Poly

pub struct Poly<R: Ring> {
    pub coeffs: Vec<R::Element>,
}

Represents a polynomial with integer coefficients.

coeffs[0] + coeffs[1]*x + coeffs[2]*x*x + ...

The coefficients should maybe be stored in reverse order.

Fields
coeffs: Vec<R::Element>
Implementations

impl<R: Ring> Poly<R>

pub const fn zero() -> Self

Returns the constant zero polynomial.

pub fn one() -> Self

Returns the constant one polynomial.

pub fn constant(c: R::Element) -> Self

Returns the constant polynomial.

pub fn from_vec(v: Vec<R::Element>) -> Self

Returns a polynomial from a list of coefficients. This function is not very intuitive because the coefficients are passed in the reverse order of what you are used to. Another reason to change the representation sometime.

pub fn random<Rand: Rng>(rng: &mut Rand, degree: usize, r: &R) -> Self

pub fn degree(&self) -> usize

Returns the degree of the polynomial. The degree of the zero polynomial is zero here!

pub fn len(&self) -> usize

Returns the number of coefficient.

pub fn is_zero(&self) -> bool

Checks whether a truncated polynomial is zero.

pub fn is_id(&self) -> bool

Checks whether a truncated polynomial is the identity polynomial.

pub fn eval(&self, a: &R::Element, r: &R) -> R::Element

Evaluate the polynomial at a using Horner’s method.

pub fn truncate(&mut self)

Truncates leading zero coefficients.

pub fn truncated(self) -> Self

Returns the truncated polynomial.

pub fn display<'a>(&'a self, var: &'a str) -> DisplayPoly<'a, R>

Returns a struct that can be used to display the polynomial.

pub fn add_assign(&mut self, rhs: &Self, ring: &R)

Add two polynomials.

pub fn add_assign_const(&mut self, rhs: &R::Element, ring: &R)

Add a constant to this polynomial.

pub fn add(&self, rhs: &Self, ring: &R) -> Self

Add two polynomials.

pub fn sub_assign(&mut self, rhs: &Self, ring: &R)

Subtract one polynomial from another.

pub fn sub_assign_const(&mut self, rhs: &R::Element, ring: &R)

Subtract a constant from the polynomial.

pub fn sub(&self, rhs: &Self, ring: &R) -> Self

Subtract one polynomial from another.

pub fn mul(&self, rhs: &Self, ring: &R) -> Self

Multiply two polynomials.

pub fn mul_assign(&mut self, rhs: &Self, ring: &R)

Multiply two polynomials. Note that this is not more efficient than Self::mul. We still allocate space for the result and do not re-use the current vector.

pub fn mul_assign_const(&mut self, rhs: &R::Element, ring: &R)

Multiply the polynomial by a constant.

pub fn mul_const(self, rhs: &R::Element, ring: &R) -> Self

Multiply the polynomial by a constant.

pub fn mul_linfac(&mut self, a: &R::Element, r: &R)

Multiplies the polynomial by a linear factor (x-a).

pub fn derivative(&self, r: &R) -> Self

Computes the derivative of the polynomial.

pub fn parse(str: String, r: &R) -> Result<Self, String>

Parse an expression from a string. This can either be a “normalized” expression:

use mba::{rings::U32, poly::Poly};
// The variable has to be called x or X.
// There may only be one term per monomial, e.g. not `3x + 4x`.
// Spaces are optional.
Poly::parse("3x^2 + 4x + 5".to_owned(), &U32).unwrap();

Or a space-separated list of coefficients a_d … a_0:

use mba::{rings::U32, poly::Poly};
Poly::parse("3 4 5".to_owned(), &U32).unwrap();

impl<R: BinaryRing> Poly<R>

pub fn shl_coeff(&mut self, m: u32, r: &R)

Shift the coefficients to the left by m.

pub fn to_expr(&self) -> ExprOp<R>

Returns an expression that uses Horner’s method to evaluate the polynomial in x.

impl<R: BinaryRing> Poly<R>

pub fn simplified(self, zi: &ZeroIdeal<R>, r: &R) -> Self

pub fn simplify(&mut self, zi: &ZeroIdeal<R>, r: &R)

Simplifies a polynomial by adding a polynomial in the zero ideal to reduce the degree of the polynomial.

pub fn reduce(&mut self, zi: &ZeroIdeal<R>, r: &R)

Reduces the degree of a polynomial if possible. This function should be used during computations where the degree could otherwise explode. It is a simplified version of the Poly::simplify algorithm, that only uses the generator of the highest degree.

impl<R: Ring> Poly<R>

pub fn to_tex(&self) -> TexPoly<'_, R>