mba

Module rings

A trait for the kind of number (ring) we are working with. These traits exist and look the way they do for multiple reasons:

So we can have extremely fast implementations for common bit widths, i.e. 8, 16, 32, 64, that use single instructions for wrapping operations and don’t have to go through an insane amount of BigInt overhead.

We can have slower but more general implementations using BigInts that allow the modulus to be set at runtime rather than compile time.

The most important trait is Ring which stores information about the ring we are working in and all operations on the ring are implemented in this trait. E.g. or U8, U16, it is empty, and for BigIntModN it contains the modulus.

An instance of Ring needs to be passed to any function that uses the ring which seems unnecessary for U8 but for BigIntModN we don’t know how to (e.g.) add elements otherwise. The other option would be to store the modulus with every element of the ring which would waste memory and we’d have to check that the moduli are equal for arithmetic operations.

The BinaryRing trait is a trait that essentially just means that we are in a ring modulo a power of two. This allows us in BinaryBigInt to just store the number of bits instead of the whole modulus. It has the bitwise operations on it.

Structs

BigIntModN

The ring of integers mod n. Each element is stored as a BigUint.

BigIntModNDataType

BinaryBigInt

The integers mod 2^n. So esentially n bit integers with wrapping semantics. The elements are stored as BigUints. The arithmetic could definitely be more efficient if num_bigint supported this use case better. We usually calculate the full result and then truncate to the number of bits, instead of ignoring all bits above the limit during the computation.

BinaryBigIntDataType

F32

Floating point numbers aren’t actually a ring (or a field) for multiple reasons:

F64

Floating point numbers aren’t actually a ring (or a field) for multiple reasons:

I8

I16

I32

I64

I128

Q

U8

U16

U32

U64

U128

VarBitsPrimitiveIntDataType

VarU8

VarU16

VarU32

VarU64

VarU128

Z

Traits

BinaryRing

Basically the integers mod a power of two.

Field

A field.

IntDivRing

A ring where you can do integer division.

OrderedRing

A ring where you can compare two elements.

Ring

A ring.

RingElement

An element of a ring. This exists mostly for convenience, so we can call e.is_zero on ring elements, but this should really all be implemented on the Ring itself, so we’d have to call r.is_zero(e).

SqrtRing

A ring where you can take approximate square roots of positive (whatever that means) elements. This is used for vectors that should have an (L2) norm. This is not really well defined but makes enough sense for all the rings it is implemented for to be useful.