mba::rings

Struct F32

pub struct F32;

Floating point numbers aren’t actually a ring (or a field) for multiple reasons:

Addition and multiplication are not associative

Infinities, NaNs, -0

But we act like it so we can store them in vectors, matrices.

Many functions assume that the floats are not NaN or infinity. Your computer won’t blow up when that happens, but some methods may return unexpected results. E.g. is_unit only checks for zero.