mba::rings

Trait IntDivRing

pub trait IntDivRing: OrderedRing {
    // Required methods
    fn rounded_div(l: &Self::Element, r: &Self::Element) -> Self::Element;
    fn euclidean_div(l: &Self::Element, r: &Self::Element) -> Self::Element;
    fn euclidean_rem(l: &Self::Element, r: &Self::Element) -> Self::Element;
}

A ring where you can do integer division.

This is needed by the crate::solver algorithms.

Required Methods

fn rounded_div(l: &Self::Element, r: &Self::Element) -> Self::Element

Divide two elements and round it to the nearest ring element.

fn euclidean_div(l: &Self::Element, r: &Self::Element) -> Self::Element

Divide two elements such that the remainder is non-negative.

fn euclidean_rem(l: &Self::Element, r: &Self::Element) -> Self::Element

Compute the remainder of the euclidean division.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.