mba::rings

Trait OrderedRing

pub trait OrderedRing: Ring {

    // Required method
    fn cmp(&self, l: &Self::Element, r: &Self::Element) -> Ordering;

    // Provided methods
    fn is_lt(&self, l: &Self::Element, r: &Self::Element) -> bool { ... }
    fn is_le(&self, l: &Self::Element, r: &Self::Element) -> bool { ... }
    fn is_ge(&self, l: &Self::Element, r: &Self::Element) -> bool { ... }
    fn is_gt(&self, l: &Self::Element, r: &Self::Element) -> bool { ... }
    fn is_positive(&self, e: &Self::Element) -> bool { ... }
    fn is_negative(&self, e: &Self::Element) -> bool { ... }
    fn abs_assign(&self, e: &mut Self::Element) { ... }
    fn abs(&self, e: Self::Element) -> Self::Element { ... }
    fn cmp_abs(&self, l: &Self::Element, r: &Self::Element) -> Ordering { ... }
    fn is_abs_eq(&self, l: &Self::Element, r: &Self::Element) -> bool { ... }
    fn is_abs_ne(&self, l: &Self::Element, r: &Self::Element) -> bool { ... }
    fn is_abs_lt(&self, l: &Self::Element, r: &Self::Element) -> bool { ... }
    fn is_abs_le(&self, l: &Self::Element, r: &Self::Element) -> bool { ... }
    fn is_abs_ge(&self, l: &Self::Element, r: &Self::Element) -> bool { ... }
    fn is_abs_gt(&self, l: &Self::Element, r: &Self::Element) -> bool { ... }

}

A ring where you can compare two elements.

This plays two roles:

In crate::solver many of the functions need OrderedRing::cmp_abs and some need OrderedRing::is_negative.

In many lattice algorithms we need to compare two elements of a crate::lattice::WorkingType, e.g. in crate::lattice::lll. And we also need OrderedRing::cmp_abs for solving systems of linear equations over a field there. For this reason this is also implemented for super::F32 and super::F64 which rust doesn’t consider to be ordered (Ord) because of NaNs. The implementation for them just unwraps the PartialOrd::partial_cmp.

Required Methods

fn cmp(&self, l: &Self::Element, r: &Self::Element) -> Ordering

Compare two elements.

Provided Methods

fn is_lt(&self, l: &Self::Element, r: &Self::Element) -> bool

Is l less than r?

fn is_le(&self, l: &Self::Element, r: &Self::Element) -> bool

Is l less than or equal to r?

fn is_ge(&self, l: &Self::Element, r: &Self::Element) -> bool

Is l greater than or equal to r?

fn is_gt(&self, l: &Self::Element, r: &Self::Element) -> bool

Is l greater than r?

fn is_positive(&self, e: &Self::Element) -> bool

Is the given element greater than 0?

fn is_negative(&self, e: &Self::Element) -> bool

Is the given element less than 0?

fn abs_assign(&self, e: &mut Self::Element)

Compute the absolute value of the element, which is defined here as is_negative(e) ? neg(e) : e.

fn abs(&self, e: Self::Element) -> Self::Element

Compute the absolute value of the element. See OrderedRing::abs_assign.

fn cmp_abs(&self, l: &Self::Element, r: &Self::Element) -> Ordering

Compare the absolute value of two elements.

fn is_abs_eq(&self, l: &Self::Element, r: &Self::Element) -> bool

Is the absolute value of l equal to the absolute value of r?

fn is_abs_ne(&self, l: &Self::Element, r: &Self::Element) -> bool

Is the absolute value of l not equal to the absolute value of r?

fn is_abs_lt(&self, l: &Self::Element, r: &Self::Element) -> bool

Is the absolute value of l less than the absolute value of r?

fn is_abs_le(&self, l: &Self::Element, r: &Self::Element) -> bool

Is the absolute value of l less than or equal to the absolute value of r?

fn is_abs_ge(&self, l: &Self::Element, r: &Self::Element) -> bool

Is the absolute value of l greater than or equal to the absolute value of r?

fn is_abs_gt(&self, l: &Self::Element, r: &Self::Element) -> bool

Is the absolute value of l greater than the absolute value of r?

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.