mba::rings

Trait Ring

pub trait Ring:
    'static
    + Clone
    + PartialEq
    + Eq
    + Debug {
    type Element: RingElement;

    // Required methods
    fn is_domain(&self) -> bool;
    fn neg_assign(&self, e: &mut Self::Element);
    fn add_assign(&self, l: &mut Self::Element, r: &Self::Element);
    fn sub_assign(&self, l: &mut Self::Element, r: &Self::Element);
    fn mul_assign(&self, l: &mut Self::Element, r: &Self::Element);
    fn is_unit(&self, e: &Self::Element) -> bool;
    fn inverse(&self, e: &Self::Element) -> Option<Self::Element>;
    fn is_zero_divisor(&self, e: &Self::Element) -> bool;
    fn random<R: Rng>(&self, rng: &mut R) -> Self::Element;
    fn element_from_usize(&self, n: usize) -> Self::Element;
    fn element_from_biguint(&self, n: &BigUint) -> Self::Element;
    fn data_type_name(&self, formatter: Formatter) -> impl Display;

    // Provided methods
    fn zero() -> Self::Element { ... }
    fn one() -> Self::Element { ... }
    fn negative_one(&self) -> Self::Element { ... }
    fn neg(&self, e: Self::Element) -> Self::Element { ... }
    fn add(&self, l: Self::Element, r: &Self::Element) -> Self::Element { ... }
    fn sub(&self, l: Self::Element, r: &Self::Element) -> Self::Element { ... }
    fn sub_assign_rhs(&self, l: &Self::Element, r: &mut Self::Element) { ... }
    fn sub_rhs(&self, l: &Self::Element, r: Self::Element) -> Self::Element { ... }
    fn mul(&self, l: Self::Element, r: &Self::Element) -> Self::Element { ... }
    fn mul_add_assign(
        &self,
        acc: &mut Self::Element,
        a: &Self::Element,
        b: &Self::Element,
    ) { ... }
    fn mul_add(
        &self,
        acc: Self::Element,
        a: &Self::Element,
        b: &Self::Element,
    ) -> Self::Element { ... }
    fn mul_sub_assign(
        &self,
        acc: &mut Self::Element,
        a: &Self::Element,
        b: &Self::Element,
    ) { ... }
    fn mul_sub(
        &self,
        acc: Self::Element,
        a: &Self::Element,
        b: &Self::Element,
    ) -> Self::Element { ... }
    fn inc_assign(&self, e: &mut Self::Element) { ... }
    fn inc(&self, e: Self::Element) -> Self::Element { ... }
    fn dec_assign(&self, e: &mut Self::Element) { ... }
    fn dec(&self, e: Self::Element) -> Self::Element { ... }
    fn square(&self, a: Self::Element) -> Self::Element { ... }
    fn element_from_bigint(&self, n: &BigInt) -> Self::Element { ... }
    fn parse_element(
        &self,
        it: &mut Peekable<impl Iterator<Item = char>>,
    ) -> Option<Self::Element> { ... }
    fn element_from_string(&self, s: &str) -> Option<Self::Element> { ... }

}

A ring.

Clone is here for convenience. All current implementations have reasonably efficient clones and deriving Clone for structs that have a R: Ring generic will not work because the R is not Clone.

Required Associated Types

type Element: RingElement

The type of the elements of the ring.

Required Methods

fn is_domain(&self) -> bool

Is the ring a domain, i.e. is 0 the only zero divisor?

fn neg_assign(&self, e: &mut Self::Element)

Negates the element in place.

fn add_assign(&self, l: &mut Self::Element, r: &Self::Element)

Add an element to another element.

fn sub_assign(&self, l: &mut Self::Element, r: &Self::Element)

Subtract one element from another.

fn mul_assign(&self, l: &mut Self::Element, r: &Self::Element)

Multiply two elements.

fn is_unit(&self, e: &Self::Element) -> bool

Check if an element is a unit. This can potentially be faster than trying to compute the inverse, so if you don’t need the inverse, you should call this.

fn inverse(&self, e: &Self::Element) -> Option<Self::Element>

Compute the multiplicative inverse of an element if it is a unit.

fn is_zero_divisor(&self, e: &Self::Element) -> bool

Check if an element is a zero divisor.

fn random<R: Rng>(&self, rng: &mut R) -> Self::Element

Returns a random element. This is used when generating values for variables that don’t have one in a valuation. The distribution is whatever you want but usually uniformly if that makes sense for the ring. This should maybe be moved into its own trait because rings whose whose elements can’t be sensibly sampled uniformly at random, e.g. Z, Q, will just panic.

fn element_from_usize(&self, n: usize) -> Self::Element

Converts the usize n into an element. The element is the result of adding 1 n-times to itself.

fn element_from_biguint(&self, n: &BigUint) -> Self::Element

Converts the BigUint n into an element. The element is the result of adding 1 n-times to itself.

fn data_type_name(&self, formatter: Formatter) -> impl Display

The name of the data type for the elements of this ring. This is used when formatting expressions. It doesn’t have to make sense.

Provided Methods

fn zero() -> Self::Element

Returns the “zero” element of the ring.

fn one() -> Self::Element

Returns the “one” element of the ring.

fn negative_one(&self) -> Self::Element

Returns the additive inverse of 1.

fn neg(&self, e: Self::Element) -> Self::Element

Negates the element, i.e. computes 0 - e.

fn add(&self, l: Self::Element, r: &Self::Element) -> Self::Element

Add two elements.

fn sub(&self, l: Self::Element, r: &Self::Element) -> Self::Element

Subtract one element from another.

fn sub_assign_rhs(&self, l: &Self::Element, r: &mut Self::Element)

Subtract an element from another element and store the result in the right-hand side.

This exists as an optimization. The naive way to do this is to use Ring::sub but that requires cloning the left-hand side. The slightly less naive way (which is the default implementation) is to rewrite the subtraction as a - b = -b + a and to then use Ring::neg_assign followed by Ring::add_assign which avoids the allocation. But not only does this function exist as a convenience function for that, but it is also possible that rewriting it like that is less optimial than a direct way, so this function can then be overriden.

Note that this isn’t required for addition, because addition is commutative.

fn sub_rhs(&self, l: &Self::Element, r: Self::Element) -> Self::Element

Subtract an element given by value from an element given by reference. See Ring::sub_assign_rhs.

fn mul(&self, l: Self::Element, r: &Self::Element) -> Self::Element

Multiply two elements.

fn mul_add_assign( &self, acc: &mut Self::Element, a: &Self::Element, b: &Self::Element, )

Multiply two elements and add the result to another element.

This exists as an optimization and has a default implementation that just allocates a new element for the product and adds that, but this can sometimes be avoided. E.g. num_bigint has a function that does this but it is not currently exposed. https://github.com/rust-num/num-bigint/blob/575cea47d21f969e541a7668751d4a82825d02bd/src/biguint/multiplication.rs#L67C37-L67C76

TODO: Expose this function and use it.

fn mul_add( &self, acc: Self::Element, a: &Self::Element, b: &Self::Element, ) -> Self::Element

Multiply two elements and add the result to another element. See Ring::mul_add_assign.

fn mul_sub_assign( &self, acc: &mut Self::Element, a: &Self::Element, b: &Self::Element, )

Ring::mul_add_assign but with Ring::sub.

fn mul_sub( &self, acc: Self::Element, a: &Self::Element, b: &Self::Element, ) -> Self::Element

Ring::mul_add but with Ring::sub.

fn inc_assign(&self, e: &mut Self::Element)

Increment the element.

fn inc(&self, e: Self::Element) -> Self::Element

Increment the element.

fn dec_assign(&self, e: &mut Self::Element)

Decrement the element.

fn dec(&self, e: Self::Element) -> Self::Element

Decrement the element.

fn square(&self, a: Self::Element) -> Self::Element

Square a number. This exists as an optimization opportunity, although no current ring takes advantage of it.

fn element_from_bigint(&self, n: &BigInt) -> Self::Element

Converts the BigInt n into an element.

If n is non-negative, this is the result of adding 1 n-times to itself.

If n is negative, this is the result of adding -1 -n-times to itself.

fn parse_element( &self, it: &mut Peekable<impl Iterator<Item = char>>, ) -> Option<Self::Element>

Parse an element from the given iterator over characters.

fn element_from_string(&self, s: &str) -> Option<Self::Element>

Parse an element from the given string.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.