mba::rings

Trait RingElement

pub trait RingElement:
    'static
    + Clone
    + PartialEq
    + Debug
    + Display {
    // Required methods
    fn zero() -> Self;
    fn is_zero(&self) -> bool;
    fn one() -> Self;
    fn is_one(&self) -> bool;
}

An element of a ring. This exists mostly for convenience, so we can call e.is_zero on ring elements, but this should really all be implemented on the Ring itself, so we’d have to call r.is_zero(e).

Required Methods

fn zero() -> Self

Returns the “zero” element of the ring.

fn is_zero(&self) -> bool

Is the given element “zero”?

fn one() -> Self

Returns the “one” element of the ring. This is already a bit scary to implement without referencing the ring because mod 1 this is 0, but we will just disallow this. (Technically RingElement::zero is as well because this is just the type we represent ring elements with e.g. BigUint and we could have BigUint::zero is the zero element in one ring but BigUint::from(42) is in another ring.)

fn is_one(&self) -> bool

Is the given element “one”?

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.