mba::rings

Trait SqrtRing

pub trait SqrtRing: Ring {
    // Required method
    fn sqrt(e: &Self::Element) -> Self::Element;
}

A ring where you can take approximate square roots of positive (whatever that means) elements. This is used for vectors that should have an (L2) norm. This is not really well defined but makes enough sense for all the rings it is implemented for to be useful.

Required Methods

fn sqrt(e: &Self::Element) -> Self::Element

Compute an approximate square root. This takes the element by reference and always allocates a new one because I don’t think there is any type that can compute the square root in-place. It also doesn’t take a self reference, because all the rings we care about don’t need it.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.