mba

Module simplify_boolean

Simplify boolean expressions using egg. There are some obvious improvements you could make:

Given a truth table, find a minimal expression using e.g. the Quine-McCluskey algorithm, before handing it to egg.

Use a custom RewriteScheduler that is aware of the cost function, i.e. tries cost reducing rewrites more aggressively.

Maybe a more refined cost function.

More refined rewrite rules.

Structs

SimplificationConfig

The way we simplify is by running equality saturation on the expression. The egraph for the expression would be prohibitively large, so that full equality saturation would take too long. egg has a lot of parameters that change the behavior of when it stops. See here for the documentation of most of these fields. When equality saturation stops without the actual saturation being complete, we start again with the current best expression. If we do not find a better expression in the next retry_same_cost iterations, we will stop. We will do this for retry_limit times and return the best expression we found.

Functions

simplify

Simplify a boolean expression.

simplify_from_truth_table

Returns a simplified boolean expression for a given truth table. The truth table is an iterator over the value of the rows. This function iterates once over the table to determine whether the DNF or the CNF is smaller and then calls simplify_from_truth_table_dnf or simplify_from_truth_table_cnf. Thus the iterator needs to be cloneable.

simplify_from_truth_table_cnf

Returns a simplified boolean expression for a given truth table. The truth table is an iterator over the indices of the rows that have a value of 0. This constructs the DNF and then simplifies it using algebra.

simplify_from_truth_table_dnf

Returns a simplified boolean expression for a given truth table. The truth table is an iterator over the indices of the rows that have a value of 1. This constructs the DNF and then simplifies it using algebra.