mba::solver

Function solve_modular_via_integer_hnf

pub fn solve_modular_via_integer_hnf(
    a: &MatrixView<Z>,
    b: &VectorView<Z>,
    n: &BigUint,
) -> AffineLattice<Z>

Solves a system of linear equations Ax=b mod n by translating it into a system of linear diophantine equations which can be solved by the standard integer HNF algorithm.

You should probably use solve_via_modular_diagonalize or solve_via_integer_diagonalize instead. solve_via_modular_diagonalize is faster for (at least) binary rings.