mba::solver

Function solve_scalar_congruence

pub fn solve_scalar_congruence<R: IntDivRing>(
    a: &R::Element,
    b: &R::Element,
    ring: &R,
) -> Option<(R::Element, R::Element)>

Solves ax=b in the ring. Returns None if there is no solution. Otherwise returns all solutions in the form (c, d) where c+di are all solutions where i ranges over all elements of the ring.

This is only meant to work for rings that are the integers mod n.