mba

Module vector

Vectors.

The reason we have a VectorStorage trait and a Vector type that just wraps the storage is so that we can implement traits for all vectors, i.e. std::ops::Index. Otherwise we would try to write something like this

impl<T, V: Vector<T>> std::ops::Index<usize> for V {
    // ...
}

which doesn’t work because V need not be defined in this crate. So we are essentially using the newtype pattern to be able to do this.

Another thing to note is that the Vectors are aware of what kind of ring the elements come from, i.e. the generic is the Ring and not the ring element. Which means you can’t have vectors of things that are not in rings. The reason is that we otherwise wouldn’t be able to implement the any functions that depend on what the ring is, e.g. arithmetic operations. Suppose the Vector didn’t take the ring as its argument but the ring element itself. We would try to implement (e.g.) add_assign like this:

impl<R: Ring, S: VectorStorage<R::Element>> Vector<R::Element, S> {
    fn add_assign<U>(&mut self, rhs: &Vector<R::Element, U>, r: &R)
        where U: VectorStorage<R::Element>
    {
        // ...
    }
}

The problem is that the type parameter R is unconstrained. In this case it feels like a limitation of rust because when we call add_assign we know which ring type R will be because we are passing an argument of that type. But if there were methods that do not take R as an argument in the impl block, then the compiler genuinely couldn’t know what R is.

Structs

OwnedVectorStorage

SliceVectorStorage

StrideStorage

This is extremely hacky, because rust currently does not support custom dynamically sized types (DST). We make this a DST by having a slice as the member, just as for ContiguousVectorStorage. We abuse the metadata of the slice to store the dimension and stride! (See std::ptr::Pointee, for what metadata is). Since we cannot implement Pointee ourselves, we have to use the usize metadata of the slice. The first 32 bits are the dimension, the second 32 bits are the stride. As I said, very hacky, but in practice the both the dimension and stride should fit into 32 bits.

StrideStorageIter

StrideStorageIterMut

Vector

A vector. See the module documentation crate::vector.

Traits

ContiguousVectorStorage

VectorStorages that lay out the elements contiguously in memory implement this.

VectorStorage

Manages how the entries of the vector are stored. See the module documentation crate::vector.

Type Aliases

OwnedVector

StrideVectorView

A vector view where elements are skipped. The stride is the amount of elements to skip not the number of bytes to skip.

VectorView

Technically this is not a vector view, because it owns the elements. However, you can only get references to this type, which are the views.