mba::vector

Struct StrideStorage

pub struct StrideStorage<R: Ring>(/* private fields */);

This is extremely hacky, because rust currently does not support custom dynamically sized types (DST). We make this a DST by having a slice as the member, just as for ContiguousVectorStorage. We abuse the metadata of the slice to store the dimension and stride! (See std::ptr::Pointee, for what metadata is). Since we cannot implement Pointee ourselves, we have to use the usize metadata of the slice. The first 32 bits are the dimension, the second 32 bits are the stride. As I said, very hacky, but in practice the both the dimension and stride should fit into 32 bits.

Implementations

impl<R: Ring> StrideStorage<R>

pub fn stride(&self) -> usize

The stride of the vector.