mba::vector

Struct Vector

pub struct Vector<R: Ring, S: VectorStorage<R> + ?Sized> { /* private fields */ }

A vector. See the module documentation crate::vector.

Implementations

impl<R: Ring, S: VectorStorage<R> + ?Sized> Vector<R, S>

pub fn from_storage(storage: S) -> Self
where S: Sized,

Construct the vector from its storage.

pub fn from_storage_ref(storage: &S) -> &Self

Creates a vector reference from a storage reference.

pub fn from_storage_ref_mut(storage: &mut S) -> &mut Self

Creates a mutable vector reference from a mutable storage reference.

pub fn dim(&self) -> usize

Returns the dimension of the vector.

pub fn entry(&self, idx: usize) -> &R::Element

Returns a reference to an entry at an index.

pub fn entry_mut(&mut self, idx: usize) -> &mut R::Element

Returns a mutable reference to an entry at an index.

pub fn iter(&self) -> S::Iter<'_>

Returns an iterator over the elements.

pub fn iter_mut(&mut self) -> S::IterMut<'_>

Returns an iterator over the mutable elements.

pub fn is_empty(&self) -> bool

Is the vector empty, i.e. dimension zero?

pub fn row_vector(&self) -> &RowVector<R, S>

Converts the vector into a row vector view, i.e. a matrix with one row.

pub fn row_vector_mut(&mut self) -> &mut RowVector<R, S>

Converts the vector into a mutable row vector view, i.e. a matrix with one row.

pub fn column_vector(&self) -> &ColumnVector<R, S>

Converts the vector into a column vector view, i.e. a matrix with one column.

pub fn column_vector_mut(&mut self) -> &mut ColumnVector<R, S>

Converts the vector into a mutable column vector view, i.e. a matrix with one column.

pub fn transform<U, F>(&self, f: F) -> OwnedVector<U>
where U: Ring, F: FnMut(&R::Element) -> U::Element,

Apply a function to each entry.

pub fn map_mut(&mut self, f: impl FnMut(&mut R::Element))

Apply a function to each entry.

pub fn swap(&mut self, i: usize, j: usize)

Swap the two entries at index i and j.

pub fn get_mut_entries( &mut self, i: usize, j: usize, ) -> (&mut R::Element, &mut R::Element)

Returns mutable references to two entries by index or panics if the indices are the same.

pub fn to<U: Ring>(&self) -> OwnedVector<U>
where R::Element: Copy + Into<U::Element>,

Convert the entries of this vector to a different type. This function is a bit restrictive because (e.g.) u32 does not implement From<&u32>. So what we do is clone the entries in the vector and then call into() on them. So because this clone is almost always unnecessary, we only implement this for Copy types.

pub fn is_zero(&self) -> bool

Are all entries in the vector zero?

pub fn neg_assign(&mut self, r: &R)

Negate the vector.

pub fn neg(self, r: &R) -> Self
where S: Sized,

Negate the vector.

pub fn add_assign<T>(&mut self, rhs: &Vector<R, T>, r: &R)
where T: VectorStorage<R> + ?Sized,

Add a vector to this one.

pub fn add_refs<T>(&self, rhs: &Vector<R, T>, r: &R) -> OwnedVector<R>
where T: VectorStorage<R> + ?Sized,

Allocate a new vector that is the sum of two vectors.

pub fn add<T>(self, rhs: &Vector<R, T>, r: &R) -> Self
where S: Sized, T: VectorStorage<R> + ?Sized,

Consumes the current vector, adds another vector to it, and returns it.

pub fn sub_assign<T>(&mut self, rhs: &Vector<R, T>, r: &R)
where T: VectorStorage<R> + ?Sized,

Subtract a vector from this one.

pub fn sub_refs<T>(&self, rhs: &Vector<R, T>, r: &R) -> OwnedVector<R>
where T: VectorStorage<R> + ?Sized,

Allocate a new vector that is the result of subtracting one vector from another.

pub fn sub<T>(self, rhs: &Vector<R, T>, r: &R) -> Self
where S: Sized, T: VectorStorage<R> + ?Sized,

Consumes the current vector, subtracts another vector from it, and returns it.

pub fn sub_assign_rhs<T>(&self, rhs: &mut Vector<R, T>, r: &R)
where T: VectorStorage<R> + ?Sized,

Subtract a vector from this one and store the result in the given vector. See also Ring::sub_assign_rhs.

pub fn sub_rhs<T>(&self, rhs: Vector<R, T>, r: &R) -> Vector<R, T>
where T: VectorStorage<R>,

Subtract a vector given by value from this one, store the result in the argument, and return it. See Self::sub_assign_rhs, Ring::sub_rhs, and Ring::sub_assign_rhs.

pub fn mul_assign(&mut self, c: &R::Element, r: &R)

Multiply the vector by a scalar.

pub fn mul(self, c: &R::Element, r: &R) -> Self
where S: Sized,

Multiply the vector by a scalar.

pub fn mul_ref(&self, c: &R::Element, r: &R) -> OwnedVector<R>

Allocate a new vector that is this vector scaled by the scalar.

pub fn mul_add_assign<T>(&mut self, c: &R::Element, v: &Vector<R, T>, r: &R)
where T: VectorStorage<R> + ?Sized,

Multiply a vector by a scalar and add the result to this vector. This exists as an optimization.

pub fn mul_add<T>(self, c: &R::Element, v: &Vector<R, T>, r: &R) -> Self
where S: Sized, T: VectorStorage<R> + ?Sized,

Multiply a vector by a scalar and add the result to this vector. This exists as an optimization.

pub fn dot<T>(&self, other: &Vector<R, T>, r: &R) -> R::Element
where T: VectorStorage<R> + ?Sized,

Compute the dot product of two vectors.

pub fn norm_sqr(&self, r: &R) -> R::Element

Computes the dot product of the vector with itself.

pub fn dist_sqr<T>(&self, other: &Vector<R, T>, r: &R) -> R::Element
where T: VectorStorage<R> + ?Sized,

Compute square euclidean distance to another vector.

pub fn norm(&self, r: &R) -> R::Element
where R: SqrtRing,

Returns the norm of the vector.

pub fn dist<T>(&self, other: &Vector<R, T>, r: &R) -> R::Element
where R: SqrtRing, T: VectorStorage<R> + ?Sized,

Computes the euclidean distance to another vector.

pub fn normalize(&mut self, r: &R)
where R: SqrtRing + Field,

Normalize a vector, i.e. divide it by its Self::norm.

pub fn reduce<T>(&mut self, m: &Matrix<R, T>, r: &R)
where R: IntDivRing, T: MatrixStorage<R> + ?Sized,

Reduce a vector using the rows a upper triangular square matrix, i.e. the first non-zero element in each row is strictly to the right of the first non-zero element in the previous row.

impl<R: Field, S: VectorStorage<R> + ?Sized> Vector<R, S>

pub fn div_assign(&mut self, c: &R::Element, r: &R)

Divide the vector by a scalar.

pub fn div(self, c: &R::Element, r: &R) -> Self
where S: Sized,

Divide the vector by a scalar.

pub fn div_ref(&self, c: &R::Element, r: &R) -> OwnedVector<R>

Allocate a new vector that is this vector divided by the scalar.

impl<R: Ring, S: ContiguousVectorStorage<R> + ?Sized> Vector<R, S>

pub fn as_slice(&self) -> &[R::Element]

Returns a slice of the vector.

pub fn as_slice_mut(&mut self) -> &mut [R::Element]

Returns a mutable slice of the vector.

impl<R: Ring> Vector<R, SliceVectorStorage<R>>

pub fn from_slice(s: &[R::Element]) -> &Self

Create a vector view of a slice.

pub fn from_slice_mut(s: &mut [R::Element]) -> &mut Self

Create a mutable vector view of a slice.

pub fn from_elem(s: &R::Element) -> &Self

Create a vector view of a single element.

pub fn from_elem_mut(s: &mut R::Element) -> &mut Self

Create a mutable vector view of a single element.

impl<R: Ring> Vector<R, OwnedVectorStorage<R>>

pub fn view(&self) -> &VectorView<R>

Returns a view of the owned vector. Currently, this has the same layout as the owned vector, so this is pretty much a no-op.

pub fn view_mut(&mut self) -> &mut VectorView<R>

Returns a mutable view of the owned vector. Currently, this has the same layout as the owned vector, so this is pretty much a no-op.

pub fn empty() -> Self

Returns an empty vector.

pub fn zero(dim: usize) -> Self

Returns a zero vector.

pub fn from_entries<U, V>(a: U) -> Self
where U: AsRef<[V]>, V: Into<R::Element> + Clone,

Creates a vector from a slice.

pub fn from_iter<U: Iterator<Item = R::Element>>(dim: usize, iter: U) -> Self

Creates a vector from an iterator.

pub fn try_from_iter<E, U: Iterator<Item = Result<R::Element, E>>>( dim: usize, iter: U, ) -> Result<Self, E>

Try to create a vector from an iterator.

pub fn from_array<U: Into<R::Element>, const D: usize>(a: [U; D]) -> Self

Creates a vector from an array.

pub fn random<Rand: Rng>(dim: usize, ring: &R, rng: &mut Rand) -> Self

Creates a random vector.

pub fn map<F: FnMut(R::Element) -> R::Element>(self, f: F) -> OwnedVector<R>

Applies a function to each entry.

pub fn append(&mut self, e: R::Element)

Appends an element to the end.

pub fn from_raw_entries(entries: Vec<R::Element>) -> Self

Returns an owned vector from a pointer and dimension.

impl<R: Ring> Vector<R, StrideStorage<R>>

pub fn from_stride_slice(s: &[R::Element], stride: usize) -> &Self

Returns a vector view with stride from a slice. The vector view will have the maximum possible dimension that fits into the slice.

pub fn from_stride_slice_mut(s: &mut [R::Element], stride: usize) -> &mut Self

Returns a vector view with stride from a slice. The vector view will have the maximum possible dimension that fits into the slice.

impl<R: Ring, S: VectorStorage<R> + ?Sized> Vector<R, S>

pub fn to_tex(&self) -> TexVector<'_, R, S>