mba::vector

Trait VectorStorage

pub trait VectorStorage<R: Ring>: 'static {
    type Iter<'a>: DoubleEndedIterator<Item = &'a R::Element>;
    type IterMut<'a>: DoubleEndedIterator<Item = &'a mut R::Element>;

    // Required methods
    fn dim(&self) -> usize;
    fn entry(&self, idx: usize) -> &R::Element;
    fn entry_mut(&mut self, idx: usize) -> &mut R::Element;
    fn iter(&self) -> Self::Iter<'_>;
    fn iter_mut(&mut self) -> Self::IterMut<'_>;
}

Manages how the entries of the vector are stored. See the module documentation crate::vector.

Required Associated Types

type Iter<'a>: DoubleEndedIterator<Item = &'a R::Element>

type IterMut<'a>: DoubleEndedIterator<Item = &'a mut R::Element>

Required Methods

fn dim(&self) -> usize

The dimension of the vector.

fn entry(&self, idx: usize) -> &R::Element

Reference to an entry at a given index.

fn entry_mut(&mut self, idx: usize) -> &mut R::Element

Mutable reference to an entry at a given index.

fn iter(&self) -> Self::Iter<'_>

Returns an iterator over the elements.

fn iter_mut(&mut self) -> Self::IterMut<'_>

Returns an iterator over the mutable elements.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.