mba::vector

Type Alias OwnedVector

pub type OwnedVector<R> = Vector<R, OwnedVectorStorage<R>>;

Aliased Type

pub struct OwnedVector<R> { /* private fields */ }

Implementations

impl<R: Ring> OwnedVector<R>

pub fn view(&self) -> &VectorView<R>

Returns a view of the owned vector. Currently, this has the same layout as the owned vector, so this is pretty much a no-op.

pub fn view_mut(&mut self) -> &mut VectorView<R>

Returns a mutable view of the owned vector. Currently, this has the same layout as the owned vector, so this is pretty much a no-op.

pub fn empty() -> Self

Returns an empty vector.

pub fn zero(dim: usize) -> Self

Returns a zero vector.

pub fn from_entries<U, V>(a: U) -> Self
where U: AsRef<[V]>, V: Into<R::Element> + Clone,

Creates a vector from a slice.

pub fn from_iter<U: Iterator<Item = R::Element>>(dim: usize, iter: U) -> Self

Creates a vector from an iterator.

pub fn try_from_iter<E, U: Iterator<Item = Result<R::Element, E>>>( dim: usize, iter: U, ) -> Result<Self, E>

Try to create a vector from an iterator.

pub fn from_array<U: Into<R::Element>, const D: usize>(a: [U; D]) -> Self

Creates a vector from an array.

pub fn random<Rand: Rng>(dim: usize, ring: &R, rng: &mut Rand) -> Self

Creates a random vector.

pub fn map<F: FnMut(R::Element) -> R::Element>(self, f: F) -> OwnedVector<R>

Applies a function to each entry.

pub fn append(&mut self, e: R::Element)

Appends an element to the end.

pub fn from_raw_entries(entries: Vec<R::Element>) -> Self

Returns an owned vector from a pointer and dimension.