mba::vector

Type Alias StrideVectorView

pub type StrideVectorView<R> = Vector<R, StrideStorage<R>>;

A vector view where elements are skipped. The stride is the amount of elements to skip not the number of bytes to skip.

Aliased Type

pub struct StrideVectorView<R> { /* private fields */ }

Implementations

impl<R: Ring> StrideVectorView<R>

pub fn from_stride_slice(s: &[R::Element], stride: usize) -> &Self

Returns a vector view with stride from a slice. The vector view will have the maximum possible dimension that fits into the slice.

pub fn from_stride_slice_mut(s: &mut [R::Element], stride: usize) -> &mut Self

Returns a vector view with stride from a slice. The vector view will have the maximum possible dimension that fits into the slice.